feat: Refactor translation management and update timeline step statuses
