+++ src/translations/index.ts
@@ -1,0 +1,68 @@
+import { TranslationType } from "../types";
+
+export const translations: Record<'en' | 'jp', TranslationType> = {
+  en: {
+    title: "4:6 Method Timer",
+    beansAmount: "Beans",
+    waterVolume: "Water",
+    taste: "Taste",
+    strength: "Strength",
+    sweet: "Sweet",
+    balance: "Balance",
+    sour: "Sour",
+    light: "Light",
+    strong: "Strong",
+    play: "Play",
+    pause: "Pause",
+    reset: "Reset",
+    flavorPour1: (amount: number) => `Pour up to ${amount}g`,
+    flavorPour2: (amount: number) => `Pour up to ${amount}g`,
+    strengthPour1: (amount: number) => `Pour up to ${amount}g`,
+    strengthPour2: (amount: number) => `Pour up to ${amount}g`,
+    strengthPour3: (amount: number) => `Pour up to ${amount}g`,
+    finish: () => "Finish",
+    language: "Language",
+    darkMode: "Dark Mode",
+    lightMode: "Light Mode",
+    roastLevel: "Roast",
+    lightRoast: "Light",
+    mediumRoast: "Medium",
+    darkRoast: "Dark",
+    waterTemp: "Temp",
+    footerMethodBy: "The 4:6 method was created by Tetsu Kasuya",
+    footerMethodVideo: "Method video",
+    footerCreatedBy: "Created by Hirokazu Takatama",
+  },
+  jp: {
+    title: "4:6メソッド タイマー",
+    beansAmount: "豆の量",
+    waterVolume: "湯量",
+    taste: "味",
+    strength: "濃さ",
+    sweet: "甘味",
+    balance: "バランス",
+    sour: "酸味",
+    light: "薄い",
+    strong: "濃い",
+    play: "再生",
+    pause: "一時停止",
+    reset: "リセット",
+    flavorPour1: (amount: number) => `${amount}g まで注湯`,
+    flavorPour2: (amount: number) => `${amount}g まで注湯`,
+    strengthPour1: (amount: number) => `${amount}g まで注湯`,
+    strengthPour2: (amount: number) => `${amount}g まで注湯`,
+    strengthPour3: (amount: number) => `${amount}g まで注湯`,
+    finish: () => "完成",
+    language: "言語",
+    darkMode: "ダークモード",
+    lightMode: "ライトモード",
+    roastLevel: "焙煎度",
+    lightRoast: "浅煎り",
+    mediumRoast: "中煎り",
+    darkRoast: "深煎り",
+    waterTemp: "湯温",
+    footerMethodBy: "4:6メソッドは粕谷哲氏によって考案されました",
+    footerMethodVideo: "メソッド解説動画",
+    footerCreatedBy: "Created by Hirokazu Takatama",
+  }
+};
